Manager_Flask/requirements.txt
aniso8601==8.0.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==7.1.2
colorama==0.4.4
cryptography==3.4.8
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
greenlet==1.1.1
gunicorn==20.1.0
idna==3.2
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
postgres==3.0.0
psycopg2-binary
psycopg2-pool
pycparser==2.20
PyJWT==2.1.0
pytz==2020.1
requests==2.26.0
six==1.15.0
SQLAlchemy==1.3.18
urllib3==1.26.6
uuid==1.30
Werkzeug==1.0.1